backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
crewai==0.1.32
langchain-anthropic==0.1.1
pymupdf==1.23.8
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
python-dotenv==1.0.0
sse-starlette==1.8.2
